Redo_main/Redo/medical_app/requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
cryptography
